requirements.txt
beautifulsoup4==4.12.3
chime==0.7.0
Pillow==10.0.0
Pillow==10.3.0
Requests==2.32.3
selenium==4.21.0
webdriver_manager==4.0.1
